backend/requirements.txt
# GAP Filler Backend Dependencies
flask>=2.0.0
flask-cors>=3.0.10
requests>=2.28.0
ollama>=0.1.0
python-dotenv>=1.0.0
reportlab>=4.0.0
